app/requirements.txt
flask
jsonify
psycopg2-binary
python-dotenv
flask_cors
openai
requests